+++ users/serializers/__init__.py
@@ -1,0 +1,51 @@
+"""Internal serializers package used to organize serializers by concern.
+
+The public import surface remains `users.serializers`.
+"""
+
+from users.serializers.profile_serializers import (
+    UserSerializer,
+    ProfileSerializer,
+    LanguageProficiencySerializer,
+)
+from users.serializers.reference_serializers import (
+    TerritorySerializer,
+    LanguageSerializer,
+    WikimediaProjectSerializer,
+    AvatarSerializer,
+    OrganizationSerializer,
+)
+from users.serializers.list_serializers import (
+    UsersBySkillSerializer,
+    UsersByTagSerializer,
+)
+from users.serializers.saved_item_serializers import (
+    EntityIdField,
+    SavedItemSerializer,
+)
+from users.serializers.badge_serializers import (
+    BadgeSerializer,
+    UserBadgeSerializer,
+)
+from users.serializers.recommendation_serializers import (
+    RecommendationUserSerializer,
+    RecommendationOrganizationSerializer,
+)
+from users.serializers.letsconnect_serializers import LetsConnectLogSerializer
+
+__all__ = [
+    # profile
+    'UserSerializer', 'ProfileSerializer', 'LanguageProficiencySerializer',
+    # reference
+    'TerritorySerializer', 'LanguageSerializer', 'WikimediaProjectSerializer', 'AvatarSerializer', 'OrganizationSerializer',
+    # lists
+    'UsersBySkillSerializer', 'UsersByTagSerializer',
+    # saved items
+    'EntityIdField', 'SavedItemSerializer',
+    # badges
+    'BadgeSerializer', 'UserBadgeSerializer',
+    # recommendations
+    'RecommendationUserSerializer', 'RecommendationOrganizationSerializer',
+    # letsconnect
+    'LetsConnectLogSerializer',
+]
